pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agential"
version = "0.0.4"
description = "A flexible agent library."
authors = ["Vincent Tu <tuvincent0106@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/alckasoc/agential"
keywords = ["agent", "LLM", "poetry"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
python = "^3.11"
faiss-cpu = "^1.7.4"
wikipedia = "^1.4.0"
func-timeout = "^4.3.5"
langchain = "^0.2.1"
tiktoken = "^0.7.0"
langchain-community = "^0.2.1"
langchain-core = "^0.2.3"
scipy = "^1.13.1"
sentence-transformers = "^2.2.2"
torch = "2.1.2"
google-api-python-client = ">=2.100.0"
ipykernel = "^6.29.5"
datasets = "^2.20.0"
plotly = "^5.22.0"
nbformat = "^5.10.4"
absl-py = "^2.1.0"
langchain-google-genai = "^1.0.10"
langchain-google-vertexai = "^1.0.10"
kaleido = "0.1.0post1"
orca = "^1.8"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
black = ">=23.9.1,<25.0.0"
ruff = "^0.0.289"
ipython = "^8.15.0"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
cookiecutter = "^2.5.0"
langchain-openai = "^0.1.8"
pandas = "1.5.3"
mypy = "^1.10.0"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests that are slow to run (deselect with '-m \"not slow\"')"
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
format = "github"
select = [
    "D",  # pydocstyle
    "I",  # isort
]
target-version = "py311"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
force-wrap-aliases = true
combine-as-imports = true
lines-between-types = 1

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
